refactor(api): extract products from client.ts
